+++ packages/application/src/ingest/index.ts
@@ -1,4 +1,5 @@
 export * from "./ports";
+export * from "./document-list-page";
 export * from "./parse-ports";
 export * from "./object-storage";
 export * from "./table-evidence";